Reworked APOD data structure
Reorganized the React component structure so that the APOD dataCache
also caches a React component with an IMG element for the associated
image. Now, data returned from APOD will be stored whole in the App
state tree, and the cached React Image component will be passed as a
prop to ImageBox.