FilmStrip viewport fix
FilmStrip will now move so that the currently selected image thumbnail
is always in view. This happens only when a new image is loaded, not
when the FilmStrip is scrolled.

Updated index.js so that the current image title is in the window
title. This makes it easier to navigate browser history.